dora-fisheye-camera/pyproject.toml
[project]
name = "dora-fisheye-camera"
version = "0.3.12"
authors = [{ name = "Xu Runtian", email = "xuruntian03@163.com" }]
description = "Dora Node for capturing video with fisheye camera"
license = "MIT"
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "dora-rs",
    "dora-rs-cli",
    "dora-rs >= 0.3.9",
    "numpy < 2.0.0",
    "opencv-python >= 4.1.1",
]
[dependency-groups]
dev = ["pytest >=8.1.1", "ruff >=0.9.1"]
[project.scripts]
dora-fisheye-camera = "dora_fisheye_camera.main:main"


[project.optional-dependencies]
# dev dependencies
dev = ["ruff", "pre-commit"]
# test dependencies
test = ["pytest"]
# doc dependencies
docs = ["sphinx"]

# ruff format and lint config
[tool.ruff]
line-length = 100
indent-width = 2
# ruff format igonres
exclude = [
    ".git", ".venv", "__pypackages__",
    "build", "dist", "node_modules"
]

[tool.ruff.lint]
# 默认启用的规则集
select = ["E4", "E7", "E9", "F"]  # 基础pycodestyle和Pyflakes规则
ignore = []
fixable = ["ALL"]  # 所有规则都可自动修复
unfixable = []
extend-select = [
  "UP",   # Ruff's UP rule
  "PERF", # Ruff's PERF rule
  "RET",  # Ruff's RET rule
  "RSE",  # Ruff's RSE rule
  "NPY",  # Ruff's NPY rule
  "N",    # Ruff's N rule
  "I",   # isort (import sorting)
  "ANN"
]

# 允许下划线前缀的未使用变量
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# 格式化风格（与Black兼容）
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true# 默认不格式化文档字符串中的代码示例

# setuptools config
[tool.setuptools.packages.find]
where = ["src"]  # 源码目录
include = [
  "dora_fisheye_camera",
  # Add other dirs
  # "other_dir",
]

# add other configs
